Logout feature implemented and sessions foundations laid
-created a logout.php file, which processes the logout and ends session

-created a global session variable: $_SESSION[‘UserName’]. If the
browser is closed, it will remember the person who is logged in, until
he logs out. (Opening an incognito tab will mean user is not signed in,
correct behaviour). Also will be used to work much of the user’s
activity.

-added a feature which prevents a user who is not logged in to be
redirected to login page (we may want to change this later, for example
you can still see a Facebook profile without being logged in at all)

-changed html file names into php, including links from home.php
footer, and settings.php logout redirect
- added a config.php document, which currently includes a global
variable for the site root (will change eventually) though, NOTE the
root folder will have to be named as ‘panoeapp’ for this to work. if we
deploy it later then it wont matter.

HTML/CSS changes

-added an example of a profile picture to the header bar (can be found
on the home.php page)